+++ engine/flink/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/sink/flink/FlinkKafkaAvroSink.scala
@@ -1,6 +1,7 @@
 package pl.touk.nussknacker.engine.avro.sink.flink
 
 import com.typesafe.scalalogging.LazyLogging
+import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import org.apache.flink.api.common.functions.{RichMapFunction, RuntimeContext}
 import org.apache.flink.formats.avro.typeutils.NkSerializableAvroSchema
 import org.apache.flink.streaming.api.datastream.DataStreamSink
@@ -23,7 +24,7 @@ class FlinkKafkaAvroSink(preparedTopic: PreparedKafkaTopic,
                          serializationSchema: KafkaSerializationSchema[KeyedValue[AnyRef, AnyRef]],
                          clientId: String,
                          // all below are passed for best effort avro encoder
-                         schema: NkSerializableAvroSchema,
+                         schema: NkSerializableAvroSchema[AvroSchema],
                          validationMode: ValidationMode)
   extends FlinkSink with Serializable with LazyLogging {
 
@@ -58,7 +59,7 @@ class FlinkKafkaAvroSink(preparedTopic: PreparedKafkaTopic,
     override def map(ctx: ValueWithContext[KeyedValue[AnyRef, AnyRef]]): KeyedValue[AnyRef, AnyRef] = {
       ctx.value.mapValue { data =>
         exceptionHandler.handling(Some(NodeComponentInfo(nodeId, "flinkKafkaAvroSink", ComponentType.Sink)), ctx.context) {
-          avroEncoder.encodeOrError(data, schema.getAvroSchema)
+          avroEncoder.encodeOrError(data, schema.getParsedSchema.rawSchema())
         }.orNull
       }
     }