+++ utils/avro-components-utils/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/SchemaRegistryProvider.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.avro.schemaregistry
 
 import cats.data.ValidatedNel
-import org.apache.avro.Schema
+import io.confluent.kafka.schemaregistry.ParsedSchema
 import pl.touk.nussknacker.engine.avro.serialization.{KafkaAvroDeserializationSchemaFactory, KafkaAvroSerializationSchemaFactory}
 import pl.touk.nussknacker.engine.kafka.RecordFormatterFactory
 
@@ -13,5 +13,5 @@ trait SchemaRegistryProvider extends Serializable with BaseSchemaRegistryProvide
 
   def recordFormatterFactory: RecordFormatterFactory
 
-  def validateSchema(schema: Schema): ValidatedNel[SchemaRegistryError, Schema]
+  def validateSchema[T <: ParsedSchema](schema: T): ValidatedNel[SchemaRegistryError, T]
 }